+++ src/app/slices/CookieConsentSlice/selectors.ts
@@ -1,0 +1,11 @@
+import { createSelector } from '@reduxjs/toolkit';
+
+import { RootState } from 'types';
+import { initialState } from '.';
+
+const selectSlice = (state: RootState) => state.cookieConsent || initialState;
+
+export const selectCookieConsent = createSelector(
+  [selectSlice],
+  state => state,
+);