+++ src/app/slices/CookieConsentSlice/types.ts
@@ -1,0 +1,6 @@
+import { CookieConsentValue } from 'app/components/OurCookieConsent';
+
+/* --- STATE --- */
+export interface CookieConsentState {
+  cookieConsentValue?: CookieConsentValue;
+}