+++ src/store/configureStore.ts
@@ -11,7 +11,7 @@ import {
   StoreEnhancer,
 } from '@reduxjs/toolkit';
 
-import { gaMiddleware } from '../utils/analytics';
+import { gaMiddlewareWithConsent } from '../utils/analytics';
 import { createReducerWithPlaceholders } from './reducers';
 import { load, saveMiddleware } from './saveload';
 
@@ -21,7 +21,7 @@ export function configureAppStore() {
   const { run: runSaga } = sagaMiddleware;
 
   // Create the store with saga middleware
-  const middlewares = [sagaMiddleware, saveMiddleware, gaMiddleware];
+  const middlewares = [sagaMiddleware, saveMiddleware, gaMiddlewareWithConsent];
 
   const preloadedState = load(true) || {};