+++ src/types/RootState.ts
@@ -3,6 +3,7 @@ import { MainPageState } from 'app/pages/MainPage/slice/types';
 import { CareerState } from 'app/slices/CareerSlice/types';
 import { ChangelogState } from 'app/slices/ChangelogSlice/types';
 import { ConnectivityState } from 'app/slices/ConnectivitySlice/types';
+import { CookieConsentState } from 'app/slices/CookieConsentSlice/types';
 import { SaveMetaState } from 'app/slices/SaveMetaSlice/types';
 import { SettingsState } from 'app/slices/SettingsSlice/types';
 import { WelcomeState } from 'app/slices/WelcomeSlice/types';
@@ -22,5 +23,6 @@ export interface RootState {
   connectivity?: ConnectivityState;
   changelog?: ChangelogState;
   settings?: SettingsState;
+  cookieConsent?: CookieConsentState;
   // [INSERT NEW REDUCER KEY ABOVE] < Needed for generating containers seamlessly
 }